pin/device burnout bugs
you can't set output high on open collector interface, as if there some static error on bus, and you would got device output low (which is allowed) and you outputting high, so you will burn out controller pin or device.
from ps/2 description:
The Data and Clock lines are both open-collector with pullup resistors to Vcc.  An "open-collector" interface has two possible state: low, or high impedance.  In the "low" state, a transistor pulls the line to ground level.  In the "high impedance" state, the interface acts as an open circuit and doesn't drive the line low or high. Furthermore, a "pullup" resistor is connected between the bus and Vcc so the bus is pulled high if none of the devices on the bus are actively pulling it low.